Support JvmCustomAnnotationValue when testing constantExpression.
When the `@Inline` annotation is used into a Xtend or DSL code, and
thefield "constantExpression" is specified into this annotation, the
Xbasecompiler fails with a class cast exception.This issue is due to the
fact that the value associated "constantExpression" is not of type
JvmBooleanAnnotationValue, but of type JvmCustomAnnotationValue (with a
XBooleanLiteral inside). This PR add unit tests.

see https://github.com/eclipse/xtext-extras/pull/324

Signed-off-by: Stéphane Galland <galland@arakhne.org>