[#397] Added validtion (and quickfixes), if ternary expression is used
If a ternary expression like "a = cond? 1 : 2" is used, an issue is raised. The default setting for this issue is ignore and can be configured in "Xtend - Errors/Warnings - Cody Style".
Also a quickfix is available (if marked as more than ignore), to replace this ternary expression with a normal if expression: "a = if (cond) 1 else 2".
This quickfix is applied recursively to nested uses of ternary expressions (like "a = cond? !cond? 1 : 2 : 3") . The issue however is raised only once fo that line.
Issue raise and Quickfix are covered with JUnitTests.

Signed-off-by: Eva Poell <epoell@itemis.com>