[#397] Added validtion, quickfixes and Java2Xtend, if ternary aka conditional expression is used.
If a ternary expression like "a = cond? 1 : 2" is used, an issue is raised. The default setting for this issue is ignore and can be configured in "Xtend - Errors/Warnings - Code Style".
Also a quickfix is available (if marked as more than ignore), to replace this ternary expression with a normal if expression: "a = if (cond) 1 else 2".
This quickfix is applied recursively to nested usages of ternary expressions (like "a = cond? !cond? 4 : 5 : 2") . The issue however is raised only once fo that line.
Issue raise and Quickfix are covered with JUnitTests.
Java2Xtend functionality: When "Paste Java Code" is clicked and if the issue is marked as ignore (= ternary expressions are allowed), a ternary expression is inserted. If the issue is marked as info/warning/error, the ternary expression will be replaces with a normal if expression. This, too, is covered with a small JUnitTest.
However this functionality only fires, if the ternary expression is not nested.

Signed-off-by: Eva Poell <epoell@itemis.com>