[#395] Support try with resources
Validate, that a given resources implements the AutoClosable Interface
and that mark as an error otherwise.

Also support content proposals for defined resources.

Finally check the whole functionality with JUnit Tests.
Both, the Error marking, as well as the compilation to correct Java Code is
checked for Java 8 and for Java versions, that do not support try with
resources.

Signed-off-by: Eva Poell <epoell@itemis.com>